Simplify code with intra-task concurrency
This is basically what I had manually done using mio, but this way
there is a single point in mainloop() where there is a single block_on
and the rest of the program is entirely single threaded boring-as-hell
Rust code.  If you look at a version using tasks then both tasks needs
to check whether the other task has already exited when "blocking" on
reads.  The overall program is actually simpler with just the single
main thread matching on which channel is currently readable because we
don’t need a stop token, etc.  The dividing line between "use tasks" &
"single task combining concurrent operations" is something along the
lines of:

1. Whether the number of tasks is known at compile time or not (and
not too many, whatever that number is)

2. Whether the tasks consume enough CPU that they may block the others
from running (assuming multi-threaded async runtime)

In this program we have 2 tasks and neither could be described as
compute intensive, so go with the simple single task multiplexing
between 2 reads approach.  This makes everything easy to follow and
the async stuff is limited to a tiny part of the code.

Note that this also updates to current smol 2.0, but the only changes
required were switching from AsRawFd to AsFd & adding the IoSafe trait
for UartRead.